Vec3 (Incomplete)

+++ Source/Engine/Renderer/Renderer.cpp
@@ -62,10 +62,10 @@ namespace kiko
 	{
 		vec2 size = texture->GetSize();
 			SDL_Rect dest;
-			dest.x = (int)x - (size.x * 0.5);
-			dest.y = (int)y - (size.y* 0.5);
-			dest.w = (int)size.x;
-			dest.h = (int)size.y;
+			dest.x = x - (size.x * 0.5);
+			dest.y = y - (size.y* 0.5);
+			dest.w = size.x;
+			dest.h = size.y;
 
 			// https://wiki.libsdl.org/SDL2/SDL_RenderCopyEx
 			SDL_RenderCopyEx(m_renderer, texture->m_texture, NULL, NULL, angle, NULL, SDL_RendererFlip::SDL_FLIP_NONE);
